requirements.txt
Click==7.1.2
Pillow==6.2.2
requests==2.22.0